Fixed "Cannot get class 'ScenePalette'." error.
Godot 3 seems to have a bug around hinting which resource type is allowed for a property if that resource was defined in GDScript.

Changed back to using the generic `Resource` type hint and setting the scene palette by default on new `SceneMap`s.

Also fixed some errors logged around connecting to the changed signal on `ScenePalette`.