Added support for rotating items (#1)
The cursor will use the currently selected axis as the axis to rotate around.  This should be a simpler workflow than trying to remember which button corresponds to the axis you want to rotate around.  The current step size is 45 degrees, but the amount can be adjusted via a constant.

Updated SceneMap itself to take an optional parameter for the rotation an item should have when placed.  This change should be backwards compatible with older versions of the addon.